feat: use the same index type in `SingleSorted2CategoryFamily` axioms
Axioms of `SingleSorted2CategoryFamily` depend on two indices, one less
than the other. Before this commit, the big index, `k`, was of type
`index` and the small type was of type `Fin k`. This was causing
problems when trying to understand indices of type `Fin k` as other
types such as `Nat` or `Fin n` in proofs.

Now, both indices are of type `index` and now the axioms take a proof
of `j < k` as an argument.