docs: add references support for the project blueprint
Since plasTeX does not call BibTeX itself when processing documents,
the printable version of the blueprint has to be run beforehand to
generate the bibliography references. For that reason, the Just recipes
for compiling the printable and the web versions of the blueprint
separatly have been made private. In this way, we enforce the use of
the `blueprint` recipe that runs both processes in the correct order.