refactor: use `LinearOrder` instead of `NatIndex` and improve naming
This commit only affects the files defining single-sorted categories and
functors. The project entry point `HigherCategoryTheory.lean` only
imports these two files for the moment. For that reason, the project
compiles. However, the rest of the files in the project still need to be
refactored to fit with these changes. So, compiling any of the other
files may lead to errors.